zfec: switch back to minimum setuptools of 0.6c3
This is newer than the one that comes with Dapper Ubuntu 6.06, *but* if there is no setuptools at all installed, then the bootstrap script will download one itself.  So with this patch, this will install cleanly on a Dapper system as long as the Dapper version of python-setuptools is *not* installed.

darcs-hash:20070430201838-eb44e-e9d102891d16eff4a510f8edd286fbc3b09c056a